# Y2018D9 huge speed-up with a DoubleLinkedList