Before

```
[day=9] (t=73ms) 375465
[day=9] (t=3698136ms) 3037741441
```

After

```
[day=9] (t=10ms) 375465
[day=9] (t=603ms) 3037741441
```